Refactor add liquidity queries into their own helper files
